requirements.txt
Django==4.0.4
django-phonenumber-field==6.1.0
